+++ src/platforms/linux/lnx_baremetal_loader.cpp
@@ -1,7 +1,7 @@
 #include <arion/arion.hpp>
 #include <arion/common/global_excepts.hpp>
-#include <arion/archs/abi_x86-64.hpp>
-#include <arion/common/abi_manager.hpp>
+#include <arion/archs/arch_x86-64.hpp>
+#include <arion/common/arch_manager.hpp>
 #include <arion/platforms/linux/lnx_baremetal_loader.hpp>
 #include <cstdint>
 #include <memory>
@@ -19,7 +19,7 @@ std::unique_ptr<LNX_LOADER_PARAMS> LinuxBaremetalLoader::process()
     if(code.size())
         params->load_address = this->map_code(code);
     
-    KERNEL_SEG_FLAGS seg_flags = arion->abi->get_attrs()->seg_flags;
+    KERNEL_SEG_FLAGS seg_flags = arion->arch->get_attrs()->seg_flags;
     if (seg_flags & ARION_VVAR_PRESENT && arion->baremetal->additional_mapped_segments.VVAR)
         params->vvar_address = this->map_vvar();
     else
@@ -73,7 +73,7 @@ ADDR LinuxBaremetalLoader::map_vvar()
     if (!arion)
         throw ExpiredWeakPtrException("Arion");
 
-    uint16_t arch_sz = arion->abi->get_attrs()->arch_sz;
+    uint16_t arch_sz = arion->arch->get_attrs()->arch_sz;
 
     ADDR vvar_load_addr = arch_sz == 64 ? LINUX_64_VVAR_ADDR : LINUX_32_VVAR_ADDR;
     ADDR vvar_sz = arch_sz == 64 ? LINUX_64_VVAR_SZ : LINUX_32_VVAR_SZ;
@@ -87,14 +87,14 @@ ADDR LinuxBaremetalLoader::map_vdso()
     if (!arion)
         throw ExpiredWeakPtrException("Arion");
 
-    uint16_t arch_sz = arion->abi->get_attrs()->arch_sz;
+    uint16_t arch_sz = arion->arch->get_attrs()->arch_sz;
 
     ADDR vdso_load_addr = arch_sz == 64 ? LINUX_64_VDSO_ADDR : LINUX_32_VDSO_ADDR;
     ADDR vdso_sz = arch_sz == 64 ? LINUX_64_VDSO_SZ : LINUX_32_VDSO_SZ;
 
     ADDR vdso_addr = arion->mem->map(vdso_load_addr, vdso_sz, LINUX_VDSO_PERMS, "[vdso]");
 
-    if (arion->abi->get_attrs()->arch == CPU_ARCH::X86_ARCH)
+    if (arion->arch->get_attrs()->arch == CPU_ARCH::X86_ARCH)
     {
         size_t bin_vdso_sz = _binary_vdso_bin_end - _binary_vdso_bin_start;
         arion->mem->write(vdso_addr, (BYTE *)_binary_vdso_bin_start, bin_vdso_sz);
@@ -108,7 +108,7 @@ ADDR LinuxBaremetalLoader::map_vsyscall()
     std::shared_ptr<Arion> arion = this->arion.lock();
     if (!arion)
         throw ExpiredWeakPtrException("Arion");
-    if (arion->abi->get_attrs()->arch != CPU_ARCH::X8664_ARCH) // For now, only x86-64 implements vsyscall segment
+    if (arion->arch->get_attrs()->arch != CPU_ARCH::X8664_ARCH) // For now, only x86-64 implements vsyscall segment
         return 0;
 
     std::array<std::string, 3> vsyscalls = {"gettimeofday", "time", "getcpu"};
@@ -121,9 +121,9 @@ ADDR LinuxBaremetalLoader::map_vsyscall()
     for (size_t syscall_i = 0; syscall_i < vsyscalls.size(); syscall_i++)
     {
         std::string syscall_name = vsyscalls.at(syscall_i);
-        uint64_t syscall_no = arion->abi->get_syscall_no_by_name(syscall_name);
+        uint64_t syscall_no = arion->arch->get_syscall_no_by_name(syscall_name);
         std::array<BYTE, VSYSCALL_ENTRY_SZ> syscall_asm =
-            static_cast<AbiManagerX8664 *>(arion->abi.get())->gen_vsyscall_entry(syscall_no);
+            static_cast<ArchManagerX8664 *>(arion->arch.get())->gen_vsyscall_entry(syscall_no);
         arion->mem->write(vsyscall_addr + syscall_i * VSYSCALL_ENTRY_SZ, syscall_asm.data(),
                           syscall_asm.size() * sizeof(BYTE));
     }