+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # 1.0.2-alpha
+- Loading coredumps
+- Memory restoring strategies
 - Implemented module for procfs
+- Changed AbiManager to ArchManager
 - Fixed unreadable symbolic link error
 
 # 1.0.1-alpha